Member 3

+++ spring/spring-demo/src/main/java/com/hana7/springdemo/jpa/controller/MemberController.java
@@ -33,7 +33,7 @@ public class MemberController {
 	}
 
 	@DeleteMapping("{id}")
-	int remove(@PathVariable Long id) {
+	long remove(@PathVariable Long id) {
 		return service.remove(id);
 	}
 }
